+++ source/voxelman/text/linebuffer.d
@@ -3,7 +3,7 @@ Copyright: Copyright (c) 2015-2017 Andrey Penechko.
 License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
 Authors: Andrey Penechko.
 */
-module voxelman.utils.linebuffer;
+module voxelman.text.linebuffer;
 
 struct LineBuffer
 {