+++ source/voxelman/text/messagewindow.d
@@ -3,11 +3,11 @@ Copyright: Copyright (c) 2015-2017 Andrey Penechko.
 License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
 Authors: Andrey Penechko.
 */
-module voxelman.utils.messagewindow;
+module voxelman.text.messagewindow;
 
 struct MessageWindow
 {
-	import voxelman.utils.linebuffer : LineBuffer;
+	import voxelman.text.linebuffer : LineBuffer;
 	LineBuffer lineBuffer;
 	alias lineBuffer this;
 	char[] inputBuf;