+++ source/voxelman/text/scale.d
@@ -3,7 +3,7 @@ Copyright: Copyright (c) 2015-2017 Andrey Penechko.
 License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
 Authors: Andrey Penechko.
 */
-module voxelman.utils.scale;
+module voxelman.text.scale;
 
 immutable string[int] scales;
 shared static this(){
@@ -25,6 +25,33 @@ shared static this(){
 	 18 : "E",
 	 21 : "Z",
 	 24 : "Y",];
+}
+
+int numDigitsInNumber(Num)(Num val)
+{
+	import std.math: abs;
+	ulong absVal = cast(ulong)abs(val);
+	int numDigits = 1;
+
+	while (absVal >= 10)
+	{
+		absVal /= 10;
+		++numDigits;
+	}
+
+	return numDigits;
+}
+
+unittest
+{
+	assert(numDigitsInNumber(0) == 1);
+	assert(numDigitsInNumber(-1) == 1);
+	assert(numDigitsInNumber(1) == 1);
+	assert(numDigitsInNumber(ubyte.max) == 3);
+	assert(numDigitsInNumber(ushort.max) == 5);
+	assert(numDigitsInNumber(uint.max) == 10); // 4294967295
+	assert(numDigitsInNumber(long.max) == 19);
+	assert(numDigitsInNumber(ulong.max) == 20);
 }
 
 int calcScale(Num)(Num val)